rajout verification si before.channel existe dans la liste, le code fonctionne maintenant

+++ klaybot.py
@@ -39,15 +39,16 @@ async def on_voice_state_update(member, before, after):
     guild = member.guild
     global clone_channel
     global list_clone
-    list_pos = list_clone.index(before.channel)
     channel = client.get_channel(685038704322281481) #On choisit notre salon textuel test2 dont l'ID est connue
     voicechannel = client.get_channel(547863579039236097) #On choisit notre salon vocal qu'on veut dupliquer
+    if before.channel in list_clone:
+        list_pos = list_clone.index(before.channel)
     if after.channel and after.channel == voicechannel and before.channel != voicechannel:
         clone_channel = await guild.create_voice_channel(f'{voicechannel.name} {member.name}', user_limit=5, category=client.get_channel(547862839369531411), position=4) #On crée un nouveau canal vocal
         list_clone.append(clone_channel)
         await channel.send(f"{member.name} a rejoint le canal {voicechannel.name} et créé le canal {clone_channel}")
         await member.edit(voice_channel=clone_channel) #On déplace le membre qui a rejoint le canal dans le nouveau canal vocal créé
-    if before.channel == list_clone[list_pos] and after.channel !=list_clone[list_pos] and len(list_clone[list_pos].members)==0:
+    if before.channel and before.channel == list_clone[list_pos] and after.channel !=list_clone[list_pos] and len(list_clone[list_pos].members)==0:
         await channel.send(f"{member.name} a quitté le canal {list_clone[list_pos]}")
         await list_clone[list_pos].delete()
         del list_clone[list_pos]
